feat(build-template-for-confirmation-email): Implement confirmation email - Install mailgenerator for confirmation email - Remove condition for generation of email for tests
