+++ package.json
@@ -37,6 +37,7 @@
     "ioredis": "^4.3.0",
     "knex": "^0.15.2",
     "lodash.isempty": "^4.4.0",
+    "mailgen": "^2.0.7",
     "nyc": "^13.1.0",
     "objection": "^1.4.0",
     "pg": "^7.6.1",
